refactor imports and fix eof test
