+++ iroh-net/src/derp/types.rs
@@ -5,8 +5,7 @@ use bytes::Bytes;
 use postcard::experimental::max_size::MaxSize;
 use serde::{Deserialize, Serialize};
 
-use super::client_conn::ClientConnManager;
-use super::PROTOCOL_VERSION;
+use super::{client_conn::ClientConnManager, codec::PROTOCOL_VERSION};
 use crate::key::PublicKey;
 
 /// A key to identify if a node belongs in a mesh
